Audit HFS+ usages of signed integers.

+++ fs/src/fs/org/jnode/fs/hfsplus/HfsPlusDirectory.java
@@ -141,7 +141,7 @@ public class HfsPlusDirectory implements FSDirectory, FSDirectoryId {
 
     @Override
     public String getDirectoryId() {
-        return Integer.toString(folder.getFolderId().getId());
+        return Long.toString(folder.getFolderId().getId());
     }
 
     @Override
